requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.10
multidict==6.0.4
openai==0.27.8
pydantic==1.10.11
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
